+++ include/kernel/syscall.h
@@ -4,10 +4,15 @@
 
 #include "arch/x86/interrupts.h"
 
+#define FD_STDIN  0u
+#define FD_STDOUT 1u
+#define FD_STDERR 2u
+
 #define SYS_PUTC  1u
 #define SYS_EXIT  2u
 #define SYS_WRITE 3u
 #define SYS_SLEEP 4u
+#define SYS_READ  5u
 
 void syscall_handler(interrupt_frame_t *frame);